backend/requirements.txt
asttokens==2.4.1
backcall==0.2.0
bcrypt==4.2.1
bidict==0.23.1
bleach==4.1.0
blinker==1.8.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.1.7
colorama==0.4.4
cryptography==35.0.0
decorator==5.1.1
docutils==0.18
exceptiongroup==1.2.2
executing==2.1.0
Faker==33.1.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.4.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
idna==3.3
importlib-metadata==4.8.1
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
keyring==23.2.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mt5linux==0.1.9
numpy==1.21.4
packaging==21.2
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pkginfo==1.7.1
pluggy==1.5.0
plumbum==1.7.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.21
Pygments==2.10.0
PyJWT==2.9.0
pyparsing==2.4.7
pytest==8.3.3
python-dateutil==2.9.0.post0
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.1
readme-renderer==30.0
requests==2.26.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rpyc==5.0.1
SecretStorage==3.3.1
simple-websocket==1.1.0
six==1.16.0
SQLAlchemy==2.0.36
stack-data==0.6.3
tomli==2.0.1
tqdm==4.62.3
traitlets==5.14.3
twine==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.7
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.6
wsproto==1.2.0
zipp==3.6.0
